test: move files to rtl-spec
